architext/requirements.txt
streamlit_vtkjs>=0.2.0
dragonfly-core>=1.32.15
ladybug-vtk>=0.8.2
honeybee-vtk==0.33.0
pollination-streamlit-io>=0.20.1
